requirements/requirements-testing.txt
coverage==6.4.1
coverage-badge==1.1.0
pytest==7.2.0
pytest-cov==3.0.0
python-dotenv